% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussie-package.R
\docType{package}
\name{ussie-package}
\alias{ussie}
\alias{ussie-package}
\title{ussie: Working with soccer data}
\description{
This is a demo package for the "Building Tidy Tools" workshop at at rstudio::conf(2022). It allows you to work with European football league data supplied by the engsoccerdata package (Curley 2016).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpflores-13/ussie}
  \item Report bugs at \url{https://github.com/jpflores-13/ussie/issues}
}

}
\author{
\strong{Maintainer}: JP Flores \email{jflores@unc.edu}

}
\keyword{internal}
